refactor:Refactor the cache package
