+++ pisces-admin/admin-api/pom.xml
@@ -15,9 +15,16 @@
     <description>Pisces Cloud 管理系统业务处理 api 模块</description>
 
     <dependencies>
+        <!-- common 核心包 -->
         <dependency>
             <groupId>com.besscroft.pisces</groupId>
             <artifactId>pisces-common</artifactId>
+            <version>${revision}</version>
+        </dependency>
+        <!-- common cache -->
+        <dependency>
+            <groupId>com.besscroft.pisces</groupId>
+            <artifactId>pisces-cache</artifactId>
             <version>${revision}</version>
         </dependency>
         <!-- openfeign依赖 1. http客户端选择okhttp 2. loadbalancer替换ribbon -->
