+++ pisces-auth/pom.xml
@@ -15,9 +15,16 @@
     <description>Pisces Cloud 认证授权中心，基于 Spring Security OAuth2</description>
 
     <dependencies>
+        <!-- common 核心包 -->
         <dependency>
             <groupId>com.besscroft.pisces</groupId>
             <artifactId>pisces-common</artifactId>
+            <version>${revision}</version>
+        </dependency>
+        <!-- common cache -->
+        <dependency>
+            <groupId>com.besscroft.pisces</groupId>
+            <artifactId>pisces-cache</artifactId>
             <version>${revision}</version>
         </dependency>
         <dependency>
