+++ pisces-framework/pisces-cache/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+  com.besscroft.pisces.framework.cache.config.RedisTemplateConfiguration
