+++ pisces-framework/pisces-common/pom.xml
@@ -19,11 +19,6 @@
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter</artifactId>
         </dependency>
-        <!-- redis -->
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-data-redis</artifactId>
-        </dependency>
         <!-- guava -->
         <dependency>
             <groupId>com.google.guava</groupId>
