--- pisces-framework/pisces-common/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  com.besscroft.pisces.framework.common.config.RedisTemplateConfiguration
