+++ pisces-framework/pom.xml
@@ -15,6 +15,7 @@
 
     <modules>
         <module>pisces-common</module>
+        <module>pisces-cache</module>
     </modules>
     <packaging>pom</packaging>
